+++ tests/data/plugins_project/libs/multiple_plugin_files/plugins/handlers.py
@@ -1,0 +1,7 @@
+from inmanta.agent.handler import ResourceHandler, provider
+from inmanta_plugins.multiple_plugin_files.helpers import helper
+
+
+@provider("std::Directory", name="myhandler")
+class MyHandler(ResourceHandler):
+    pass